refactor(mon-pix): rename campaign results route
