+++ mon-pix/app/components/campaigns/assessment/results/evaluation-results-hero/acquired-badges.gjs
