+++ mon-pix/app/components/campaigns/assessment/results/evaluation-results-hero/custom-organization-block.gjs
