+++ mon-pix/app/components/campaigns/assessment/results/evaluation-results-tabs/results-details/index.gjs
