+++ mon-pix/app/components/campaigns/assessment/results/evaluation-results-tabs/rewards/badge.gjs
