+++ mon-pix/app/styles/app.scss
@@ -173,7 +173,6 @@ of an adaptative/mobile-first approach — refactoring is welcome here */
 @import 'pages/account-recovery';
 @import 'pages/send-profile';
 @import 'pages/sign-form';
-@import 'pages/skill-review';
 @import 'pages/sso-selection';
 @import 'pages/terms-of-service';
 @import 'pages/update-expired-password-form';
