+++ mon-pix/app/styles/components/campaigns/assessment/results/evaluation-results-tabs/rewards.scss
