+++ mon-pix/app/styles/components/campaigns/assessment/results/quit-results.scss
