+++ mon-pix/app/styles/components/campaigns/index.scss
@@ -1,6 +1,5 @@
-@import 'assessment/skill-review/quit-results';
-@import 'assessment/skill-review/share-badge-icons';
-@import 'assessment/skill-review/evaluation-results-tabs/results-details';
-@import 'assessment/skill-review/evaluation-results-tabs/rewards';
-@import 'assessment/skill-review/evaluation-results-tabs/trainings';
+@import 'assessment/results/quit-results';
+@import 'assessment/results/evaluation-results-tabs/results-details';
+@import 'assessment/results/evaluation-results-tabs/rewards';
+@import 'assessment/results/evaluation-results-tabs/trainings';
 @import 'invited/learner-reconciliation';
