+++ mon-pix/app/styles/pages/_evaluation-results.scss
@@ -231,7 +231,7 @@
   gap: var(--pix-spacing-10x);
 
   &::after {
-    background-image: url("/images/illustrations/skill-review/retry-assessment.svg");
+    background-image: url("/images/illustrations/results/retry-assessment.svg");
     background-repeat: no-repeat;
     background-position: center;
     background-size: contain;
