+++ mon-pix/app/templates/campaigns/assessment/results.hbs
@@ -1,0 +1,3 @@
+{{page-title (t "pages.skill-review.title")}}
+
+<Routes::Campaigns::Assessment::EvaluationResults @model={{@model}} />
