+++ mon-pix/tests/integration/components/campaigns/assessment/results/results-details/results-details-test.js
@@ -33,9 +33,7 @@ module('Integration | Components | Campaigns | Assessment | EvaluationResultsTab
 
     // when
     screen = await render(
-      hbs`<Campaigns::Assessment::SkillReview::EvaluationResultsTabs::ResultsDetails
-  @competenceResults={{this.competenceResults}}
-/>`,
+      hbs`<Campaigns::Assessment::Results::EvaluationResultsTabs::ResultsDetails @competenceResults={{this.competenceResults}} />`,
     );
   });
